+++ react-app/src/App.js
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
-import LoginForm from './components/auth/LoginForm';
-import SignUpForm from './components/auth/SignUpForm';
+import LoginForm from './components/LoginPage/LoginForm';
+import SignUpForm from './components/SignupPage/SignUpForm';
 import ProtectedRoute from './components/auth/ProtectedRoute';
 import UsersList from './components/UsersList';
 import { authenticate } from './store/session';
@@ -35,7 +35,7 @@ function App() {
 					<LoginForm />
 				</Route>
 
-				<Route path='/sign-up' exact={true}>
+				<Route path='/signup' exact={true}>
 					<Navbar />
 					<SignUpForm />
 				</Route>
